+++ cpu.h
@@ -3,7 +3,7 @@
 #include "comm.h"
 
 
-extern uint32_t speed_multiplier;
+extern double speed_multiplier;
 
 void reset_cpu_states();
 void cpu_run();
